feat: enable comprehensive warnings and errors for GNU compiler; fix return type in integer_pack and add virtual destructor in lie_potential

+++ include/mstd/pack/integer_pack.hpp
@@ -93,7 +93,7 @@ namespace mstd
                 if (vals[i] != 0)
                     return i;
 
-            return -1;
+            return static_cast<size_t>(-1);
         }
     };
